adding TransactionCreationException and ZeroAmountTransactionException also adjusted create_transaction.html

+++ src/main/java/com/company/web/wallet/services/TransactionServiceImpl.java
@@ -1,14 +1,20 @@
 package com.company.web.wallet.services;
 
 import com.company.web.wallet.exceptions.AuthorizationException;
+import com.company.web.wallet.exceptions.ZeroAmountTransactionException;
 import com.company.web.wallet.helpers.TransactionType;
 import com.company.web.wallet.models.Transaction;
 import com.company.web.wallet.models.User;
 import com.company.web.wallet.repositories.TransactionRepository;
+import com.company.web.wallet.repositories.TransactionRepositoryImpl;
 import com.company.web.wallet.repositories.UserRepository;
+import com.company.web.wallet.repositories.UserRepositoryImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.Comparator;
@@ -20,6 +26,7 @@ public class TransactionServiceImpl implements TransactionService {
     private final TransactionRepository transactionRepository;
     private final WalletService walletService;
     private final UserRepository userRepository;
+    private final Logger logger = LoggerFactory.getLogger(TransactionRepositoryImpl.class);
 
 
     @Autowired
@@ -28,6 +35,7 @@ public class TransactionServiceImpl implements TransactionService {
         this.walletService = walletService;
         this.userRepository = userRepository;
     }
+
     @Override
     public List<Transaction> getTransactions(
             String username,
@@ -76,43 +84,59 @@ public class TransactionServiceImpl implements TransactionService {
     public List<Transaction> getTransactionsByUser(User user) {
         return transactionRepository.getTransactionsByUser(user);
     }
+
     @Override
     public List<Transaction> getTransactionsByType(TransactionType type) {
         return transactionRepository.getTransactionsByType(type);
     }
+
     @Override
     public List<Transaction> getTransactionsByDateRange(LocalDateTime startDate,
                                                         LocalDateTime endDate) {
         return transactionRepository.getTransactionsByDateRange(startDate, endDate);
     }
+
     @Override
     public List<Transaction> getTransactionsBySender(User authenticatedUser, User sender) {
 
         checkAccessPermissions(authenticatedUser, sender.getId());
         return transactionRepository.getTransactionsBySender(sender);
     }
+
     @Override
     public List<Transaction> getTransactionsByRecipient(User authenticatedUser, User recipient) {
         checkAccessPermissions(authenticatedUser, recipient.getId());
         return transactionRepository.getTransactionsByRecipient(recipient);
     }
+
     @Override
     public List<Transaction> getTransactionsByDirection(TransactionType direction) {
         return transactionRepository.getTransactionsByDirection(direction);
     }
+
     @Override
     public void createTransaction(Transaction transaction) {
-        User sender = userRepository.getByUsername(transaction.getSender().getUsername());
-        User recipient = userRepository.getByUsername(transaction.getRecipient().getUsername());
-        transaction.setSender(sender);
-        transaction.setRecipient(recipient);
-        walletService.addToBalance(walletService.getWalletIdForUser(transaction.getRecipient()),
-                transaction.getRecipient(), transaction.getAmount());
-        walletService.removeFromBalance(walletService.getWalletIdForUser(userRepository.getByUsername(transaction.getSender().getUsername())),
-                transaction.getSender(),
-                transaction.getAmount());
-        transactionRepository.createTransaction(transaction);
+        try {
+            if (transaction.getAmount().compareTo(BigDecimal.ZERO) <= 0) {
+                logger.error("ZeroAmountTransactionException: Transaction amount must be greater than zero.");
+                return;
+            }
+            User sender = userRepository.getByUsername(transaction.getSender().getUsername());
+            User recipient = userRepository.getByUsername(transaction.getRecipient().getUsername());
+            transaction.setSender(sender);
+            transaction.setRecipient(recipient);
+            walletService.addToBalance(walletService.getWalletIdForUser(transaction.getRecipient()),
+                    transaction.getRecipient(), transaction.getAmount());
+            walletService.removeFromBalance(walletService.getWalletIdForUser(userRepository.getByUsername(transaction.getSender().getUsername())),
+                    transaction.getSender(),
+                    transaction.getAmount());
+            transactionRepository.createTransaction(transaction);
+        } catch (RuntimeException e) {
+            throw new RuntimeException(e.getMessage());
+        }
+
     }
+
     @Override
     public void deleteTransaction(Long id) {
         transactionRepository.deleteTransaction(id);
